requirements.txt
pip==8.1.2
blinker==1.4
cffi==1.7.0
click==6.3
contextlib2==0.5.4
cryptography==1.4
dateparser==0.3.4
enum==0.4.6
enum34==1.1.2
Flask==0.10.1
functools32==3.2.3.post2
future==0.15.2
gunicorn==19.3.0
honcho==0.6.6
idna==2.1
ipaddress==1.0.16
itsdangerous==0.24
jdatetime==1.7.4
Jinja2==2.8
jsonschema==2.5.1
jsontableschema==0.5.1
MarkupSafe==0.23
ndg-httpsclient==0.4.2
ordereddict==1.1
paramiko==2.0.2
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pynYNAB
-e git+https://github.com/scottrobertson/nYNABapi.git@master#egg=pynYNAB

pyOpenSSL==16.0.0
python-dateutil==2.5.0
python-dotenv==0.5.1
pytz==2015.7
PyYAML==3.11
raven==5.23.0
regex==2016.3.2
rfc3987==1.3.5
schematics==2.0.0.dev2
six==1.10.0
toolz==0.7.4
umalqurra==0.2
unicodecsv==0.14.1
Werkzeug==0.11.5
sqlalchemy
configargparse
configparser
appdirs
requests
aenum
lxml
-e git://github.com/csingley/ofxtools.git@4a3e496d7d2a3babdeabd4e114d98fc5f2862667#egg=ofxtools
pynab